src/FolkerKinzel.VCards.Tests/BuilderParts/IDBuilderTests.cs
namespace FolkerKinzel.VCards.BuilderParts.Tests;

[TestClass]
public class IDBuilderTests
{
    [TestMethod]
    [ExpectedException(typeof(InvalidOperationException))]
    public void SetTest1() => new IDBuilder().Set(null);

    [TestMethod]
    [ExpectedException(typeof(InvalidOperationException))]
    public void ClearTest1() => new IDBuilder().Clear();

    [TestMethod]
    public void EqualsTest1()
    {
        Assert.IsFalse(new IDBuilder().Equals((IDBuilder?)null));

        var builder = new IDBuilder();
        Assert.AreEqual(builder.GetHashCode(), ((object)builder).GetHashCode());
    }

    [TestMethod]
    public void ToStringTest1() => Assert.IsNotNull(new IDBuilder().ToString());
}
